fix: interfaces fixes
